requirements.txt
qrcode
pillow
requests
